Disable action bar for Espresso tests as much as possible
The action bar overlap the button used by EspressoTest,
and the button's click operation will fail/flaky.
This CL uses NoActionBar theme for EspressoActivity
to avoid this flaky.
