+++ integration_tests/androidx_test/src/main/AndroidManifest-NoTestPackageActivities.xml
@@ -12,7 +12,12 @@
         <activity android:name="org.robolectric.integrationtests.axt.EspressoActivity"
                   android:label="Activity Label"
                   android:exported="true" />
+        <activity android:name="org.robolectric.integrationtests.axt.NoActionBarEspressoActivity"
+            android:label="Activity Label"
+            android:theme="@style/Theme.AppCompat.NoActionBar"
+            android:exported="true" />
         <activity android:name="org.robolectric.integrationtests.axt.EspressoScrollingActivity"
+            android:theme="@style/Theme.AppCompat.NoActionBar"
             android:exported="true" />
         <activity android:name="org.robolectric.integrationtests.axt.ActivityWithPlatformMenu"
             android:exported="true">
