+++ integration_tests/androidx_test/src/main/java/org/robolectric/integrationtests/axt/NoActionBarEspressoActivity.java
@@ -1,0 +1,8 @@
+package org.robolectric.integrationtests.axt;
+
+/**
+ * The test {@link android.app.Activity} for Espresso testing but without action bar.
+ *
+ * <p>It is used for Espresso tests that don't want action bar affect its content.
+ */
+public class NoActionBarEspressoActivity extends EspressoActivity {}
